+++ client/src/App/index.js
@@ -24,7 +24,6 @@ import MyFoodPage from '../Pages/MyFood';
 import FoodPage from '../Pages/Food';
 import ExercisePage from '../Pages/Exercise';
 import Provider from '../components/userProvider';
-import Footer from '../components/Footer';
 import DesktopNavBar from '../components/DesktopNavBar';
 import './style.css';
